requirements.txt
# django-storages-azure==1.6.8
appdirs==1.4.3
APScheduler~=3.9.1
astroid==2.3.1
attrs==19.3.0
autopep8==1.4.4
azure-common==1.1.25
azure-core==1.8.2
azure-functions==1.5.0
azure-identity==1.5.0
azure-keyvault-secrets==4.2.0
azure-servicebus==7.0.1
azure-storage-blob==12.5.0
azure-storage-common==2.1.0
bandit==1.6.2
cached-property==1.5.1
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
defusedxml==0.6.0
django-bulk-update==2.2.0
django-cors-headers==2.5.3
django-extensions==2.2.9
django-filter==2.2.0
django-pyodbc-azure==2.1.0.0
django-rest-swagger==2.2.0
Django==2.1.13
djangorestframework-api-key==2.0.0
djangorestframework==3.10.3
djangorestframework-api-key==2.0.0
entrypoints==0.3
flake8==3.7.8
flower==0.9.5
gitdb2==2.0.6
GitPython==3.0.3
holidays==0.11.1
idna==2.8
ipython==7.24.1
isodate==0.6.0
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
jsonschema==4.4.0
lazy-object-proxy==1.4.2
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
model-bakery==1.2.0
oauthlib==3.1.0
openapi-codec==1.3.2
openpyxl~=3.0.5
pandas==1.2.4
pbr==5.4.3
pycodestyle==2.5.0
pyflakes==2.1.1
pylint-django==2.0.11
pylint-plugin-utils==0.6
pylint==2.4.2
pyodbc==4.0.30
python-decouple==3.4.0
pytz==2019.3
PyYAML==5.1.2
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
smmap2==2.0.5
stevedore==1.31.0
termcolor==1.1.0
tqdm==4.36.1
typed-ast==1.5.1
uritemplate==3.0.0
urllib3==1.25.6
Werkzeug
whitenoise==4.1.4
wrapt==1.11.2
XlsxWriter~=1.3.7
xmlrunner==1.7.7
zeep==3.4.0
reportlab~=3.5.67
# django-storages-azure==1.6.8
